move theme code into namespace
