+++ LookupAnything/Framework/ModConfig.cs
@@ -1,6 +1,7 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.Serialization;
 using Pathoschild.Stardew.Common;
+using Pathoschild.Stardew.LookupAnything.Framework.Themes;
 
 namespace Pathoschild.Stardew.LookupAnything.Framework;
